backend/requirements.txt
aioredis==2.0.1
amqp==5.1.1
apns2==0.7.2
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.12.1
billiard==4.1.0
boto3==1.28.30
botocore==1.31.30
CacheControl==0.13.1
cachetools==5.3.1
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==41.0.3
cssselect2==0.7.0
daphne==4.0.0
decorator==4.4.2
Django==3.2
django-common-helpers==0.9.2
django-cors-headers==4.2.0
django-cron==0.6.0
django-crontab==0.7.1
django-easy-pdf==0.1.1
django-push-notifications==3.0.0
django-rest-swagger==2.2.0
django-simple-history==3.3.0
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
fcm-django==2.0.0
firebase-admin==6.0.0
flower==2.0.1
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
h2==2.6.2
hiredis==2.2.3
hpack==3.0.0
html2text==2020.1.16
html5lib==1.1
http-ece==1.1.0
httplib2==0.22.0
humanize==4.8.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==21.0.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.1
lxml==4.9.3
MarkupSafe==2.1.3
moviepy==1.0.3
msgpack==1.0.5
numpy==1.25.2
openapi-codec==1.3.2
oscrypto==1.3.0
pdfkit==1.0.0
Pillow==10.0.0
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.1
psycopg2-binary==2.9.7
py-vapid==1.9.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyfcm==1.5.4
PyHamcrest==2.0.4
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
PyJWT==1.7.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypdf==3.15.1
PyPDF2==3.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.3.7
pytz==2023.3
pywebpush==1.14.0
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.0
reportlab==3.6.13
requests==2.31.0
rsa==4.9
s3transfer==0.6.2
sendgrid==3.6.5
sendgrid-django==4.2.0
service-identity==23.1.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
stripe==6.0.0
svglib==1.5.1
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
xhtml2pdf==0.2.11
zipp==3.16.2
zope.interface==6.0
